Add files via upload
This is an menu driven calculator, in which an menu card is given to user and based on user choice(input) the operation is done, and the user can do operation till he didn't exit from code. This is done by creating multiple functions for each operation and menu of operations is build using while loop. This is my First project built using python basics.