Add files via upload
This is an number guessing game, in which user have to guess number between 1 to 100, and when user enter the number code tell user that is he close to answer to away from answer. And when user guess the write number then Congratulation message is display with attempt use to guess the number. This i have build using random library and in it i have use while loop and conditional statements. This was challenging for me, as i have not fully learnt python and its libraries, but based on the basic knowledge i created it.